moved cms reducers and actions to separate files.
