+++ client/src/actions/CMSActions.js
@@ -1,0 +1,14 @@
+import { CREATE_ITEM } from './types';
+
+export const createItem = (name,desc,image,price,type) =>{
+    return{
+      type: CREATE_ITEM,
+      payload: {
+      name,
+      desc,
+      image,
+      price,
+      type
+      }
+    }
+}
