+++ client/src/actions/index.js
@@ -1,1 +1,3 @@
+export * from './login';
+export * from './CMSActions';
 export * from './OrderItemActions';
