+++ client/src/actions/login.js
@@ -1,44 +1,37 @@
 import axios from 'axios';
 import { AsyncStorage } from 'react-native';
 import { Actions, ActionConst } from 'react-native-router-flux';
-// import Config from 'react-native-config';
-
-export const LOGIN_REQUEST = 'LOGIN_REQUEST';
-export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
-export const LOGIN_FAILURE = 'LOGIN_FAILURE';
-export const LOGOUT_REQUEST = 'LOGOUT_REQUEST';
-export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-export const LOGOUT_FAILURE = 'LOGOUT_FAILURE';
+import { LOGIN_REQUEST, LOGIN_SUCCESS, LOGIN_FAILURE, LOGOUT_REQUEST, LOGOUT_SUCCESS, LOGOUT_FAILURE } from './types';
 
 const requestLogin = creds => ({
-    type: 'LOGIN_REQUEST',
+    type: LOGIN_REQUEST,
     isFetching: true,
     isAuthenticated: false,
     creds,
 });
 
 const receiveLogin = user => ({
-    type: 'LOGIN_SUCCESS',
+    type: LOGIN_SUCCESS,
     isFetching: false,
     isAuthenticated: true,
     id_token: user.id_token,
 });
 
 const loginError = message => ({
-    type: 'LOGIN_FAILURE',
+    type: LOGIN_FAILURE,
     isFetching: false,
     isAuthenticated: false,
     message,
 });
 
 const requestLogout = () => ({
-    type: 'LOGOUT_REQUEST',
+    type: LOGOUT_REQUEST,
     isFetching: true,
     isAuthenticated: true,
 });
 
 const receiveLogout = () => ({
-    type: 'LOGOUT_SUCCESS',
+    type: LOGOUT_SUCCESS,
     isFetching: false,
     isAuthenticated: false,
 });
