--- client/src/reducers/AnimationReducers
@@ -1,16 +1,0 @@
-import { EMPLOYEE_UPDATE, EMPLOYEE_CREATE } from '../actions/types';
-
-const INITIAL_STATE = {
-  name: '',
-  phone: '',
-  shift: ''
-};
-
-export default (state = INITIAL_STATE, action) => {
-  switch (action.type) {
-    case EMPLOYEE_UPDATE:
-      return { ...state, [action.payload.prop]: [action.payload.value]  }; //key interpolation, not array
-    default:
-      return state;
-  }
-};
