+++ client/src/reducers/index.js
@@ -1,12 +1,12 @@
 import { combineReducers } from 'redux';
-// import Settings from '../components/admin/Settings'
-// import SettingReducers from './SettingReducers';
+import cms__reducer from './cms__reducer';
 import item__reducer from './item__reducer';
 import authReducer from './authReducer';
 
 const RootReducer = combineReducers({
   auth: authReducer,
-  items:item__reducer
+  cms: cms__reducer,
+  items:item__reducer,
 });
 
 export default RootReducer;
